MAJOR: Using BigDecimal and BigInteger instead of double and long for precision and to prevent overflows.

+++ src/main/java/com/ahaviss/calculators/geocalc/shapes2D/Shape2D.java
@@ -1,5 +1,7 @@
 package com.ahaviss.calculators.geocalc.shapes2D;
 
+import java.math.BigDecimal;
+
 public abstract class Shape2D {
-    public abstract double area();
+    public abstract BigDecimal area();
 }
