+++ src/main/java/com/ahaviss/calculators/geocalc/shapes3D/Shape3D.java
@@ -1,6 +1,8 @@
 package com.ahaviss.calculators.geocalc.shapes3D;
 
+import java.math.BigDecimal;
+
 public abstract class Shape3D {
-    public abstract double volume();
-    public abstract double surfaceArea();
+    public abstract BigDecimal volume();
+    public abstract BigDecimal surfaceArea();
 }
